requirements.txt
flatbuffers==2.0
pyzmq==22.2.1
gpiozero==1.6.2
